refactor(shim): Convert to a shim
After exploring CoreObject more as we're trying to fix ES Classes in
general, we've found a way to monkeypatch Ember instead of using a babel
transform. This method is likely to be much more efficient and inline
with future behavior of classes, so we're opting to switch to this
strategy.